Squashed commit of the following:
commit b1700b23047905881e98f91817b877e548e4f77c
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:50:28 2024 +0530

    fix

commit 127955535f27aed1b9d04bc4fa57d9b0e6497e39
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:38:58 2024 +0530

    in previous commit I tested by changing compiler to gcc, still fails in macOS

commit 5920f1a891f64bd9030737d2804103f12246899d
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:28:13 2024 +0530

    .

commit 81918f1d6bf626db1cf838884d6536c137d735f7
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:19:22 2024 +0530

    .

commit f8ed783f3c7dbef6eba899de587f8d64cad916dc
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:12:23 2024 +0530

    .

commit 8f709d1d26c74c4d69bd5e9f26951819e217c3f8
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:05:16 2024 +0530

    .

commit 69b50d1c7087baa961241777cb29819fafb00568
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 15:01:03 2024 +0530

    .

commit 797f644b9307de9b72c6eb7b7eae71a3028e8d7c
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 07:45:37 2024 +0530

    fix .

commit 7e1367349438e4599133649675127cd975f28dc7
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 07:37:19 2024 +0530

    fix .

commit e18a15c7ba441f8022ed32432825c197ff279130
Author: Vipul Cariappa <vipulcariappa@gmail.com>
Date:   Mon Jul 8 07:34:19 2024 +0530

    CI fix

commit de257f59147f71302dd2eb7c7479768bbde2b2c4
Merge: 9ec6d6de4 f414b0f05
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 15:03:16 2024 -0600

    Merge branch 'main' into ci_mamba

commit 9ec6d6de420cb46c314cf0f40e6a17d63bde8fd9
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 14:58:15 2024 -0600

    X

commit e69e069b208b732e57aeb6da0ef70b8651b9e217
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 14:44:51 2024 -0600

    Pin zstd, zlib

commit 2ed3ac2dec1b87be26d1f0996cc15fd13484be81
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 14:36:39 2024 -0600

    downgrade

commit c8a80c3ec97bad027c539928dd0ddfb3ca33e076
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 12:16:48 2024 -0600

    zstd

commit 1c77ac52adbb5603c3e3ad4c876528d302c1de96
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 12:14:14 2024 -0600

    X

commit ba8a13ca3ffb5b04d5114f8fba2a38ba3e4471d4
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 12:12:20 2024 -0600

    Fix nlohmann_json

commit 81beb7c9e372a2b53d0a87500df6f8837b68d3e0
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 12:10:20 2024 -0600

    Pin packages

commit e4fa09a6329cff6430cace0b016f49d7fc71aaa0
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:58:09 2024 -0600

    List

commit ae166560354fb3b5e395fc471e9422e644f6a441
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:54:47 2024 -0600

    Do not upgrade installed packages

commit e73253e4e4b17f0c3d874b3b17cc1876a2aa3bc2
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:50:47 2024 -0600

    Pin the packages

commit d1dc5f0c9102ab882e0df59fd75b19e0859b27ad
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:47:13 2024 -0600

    Use latest version

commit 0c3b95b491a87507eafbe33947a49fb55479a2ab
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:39:46 2024 -0600

    Fix Windows test

commit 34613d7c60e251bce349fb3f52c145f2af929d88
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:24:17 2024 -0600

    Use xonsh 0.16.0

commit d1e230a7f27883564f9e5544120612e84f5685a0
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:22:01 2024 -0600

    Use the old xeus

commit eed485e2c807f610afbc1ebd75d2f1b4a6ab9b51
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:19:56 2024 -0600

    Use lp

commit e6574c4a9db8a0e4bee5e14a1cc52c865f56b096
Author: Ondřej Čertík <ondrej@certik.us>
Date:   Sat Jul 6 11:14:37 2024 -0600

    CI: switch to micromamba, update xeus

+++ src/lpython/python_evaluator.cpp
@@ -143,11 +143,11 @@ Result<PythonCompiler::EvalResult> PythonCompiler::evaluate(
                                     ->m_symtab->get_symbol(run_fn);
             LCOMPILERS_ASSERT(fn)
             if (ASRUtils::get_FunctionType(fn)->m_return_var_type->type == ASR::ttypeType::UnsignedInteger) {
-                uint8_t r = e->execfn<uint8_t>(run_fn);
+                uint8_t r = e->execfn<int>(run_fn);
                 result.type = EvalResult::unsignedInteger1;
                 result.u32 = r;
             } else {
-                int8_t r = e->execfn<int8_t>(run_fn);
+                int8_t r = e->execfn<int>(run_fn);
                 result.type = EvalResult::integer1;
                 result.i32 = r;
             }
@@ -156,11 +156,11 @@ Result<PythonCompiler::EvalResult> PythonCompiler::evaluate(
                                     ->m_symtab->get_symbol(run_fn);
             LCOMPILERS_ASSERT(fn)
             if (ASRUtils::get_FunctionType(fn)->m_return_var_type->type == ASR::ttypeType::UnsignedInteger) {
-                uint16_t r = e->execfn<uint16_t>(run_fn);
+                uint16_t r = e->execfn<int>(run_fn);
                 result.type = EvalResult::unsignedInteger2;
                 result.u32 = r;
             } else {
-                int16_t r = e->execfn<int16_t>(run_fn);
+                int16_t r = e->execfn<int>(run_fn);
                 result.type = EvalResult::integer2;
                 result.i32 = r;
             }
